pyproject.toml
[project]
name = "jons-mcp-file-server"
version = "0.1.0"
description = "File server abstraction for MCP servers - supports GCS signed URLs and localhost"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-storage>=2.0.0",
    "pyngrok>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jons_mcp_file_server"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
